Fix the correctness check problem
Problem was caused by the way errorCells were deleted. When user adds an value in to the grid, every error cell is checked if the change above haven't made them valid.

Solution: instead of bools for storing the information, if certain value can be validly used in a given cell, are UsageCounters used.